requirements.txt
# Core Django
Django==4.2.7
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# Database
psycopg2-binary==2.9.9  # For PostgreSQL (optional)

# AI and Image Processing
Pillow==10.1.0
google-generativeai==0.3.2  # For Gemini AI
openai==1.3.0  # For OpenAI (optional)

# Utilities
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0

# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Production
gunicorn==21.2.0  # For Linux/macOS
waitress==3.0.0   # For Windows

# Data Processing
numpy==1.24.3
pandas==2.1.3

# File Watching (Optional)
watchdog==3.0.0